#7126 export shouldPatch to helper functions #7126 small optimizations
