+++ frontend/src/components/app/OverlayField.js
@@ -216,7 +216,7 @@ class OverlayField extends Component {
     return (
       <div
         className="overlay-field"
-        onKeyDown={(e) => this.handleKeyDown(e)}
+        onKeyDown={this.handleKeyDown}
         tabIndex={-1}
       >
         {renderedContent}
